requirements.txt
Flask==3.1.3
requests==2.32.3
python-dotenv==1.2.2
notion-client==3.0.0
openai>=1.0.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
